Small cleanups
Removed a test that turned out to not be useful in a feasible fashion (not worth effort and
also required some ugly modifications to the code just for the purposes of testing that
actually lowered the overall readibility and quality of the code).

Cleaned up an input param for the loadPages method of the pdf viewer, and also fixed up
the pdfjs lib mock to return the correct value.